+++ node/index.ts
@@ -2,6 +2,8 @@ import type { ClientsConfig, ServiceContext, RecorderState } from '@vtex/api'
 import { method, Service } from '@vtex/api'
 
 import { Clients } from './clients'
+import { parseFile } from './middlewares/notify/parseFile'
+import { startEventChain } from './middlewares/notify/startEventChain'
 
 const TIMEOUT_MS = 800
 
@@ -19,15 +21,16 @@ declare global {
   type Context = ServiceContext<Clients, State>
 
   interface State extends RecorderState {
-    code: number
+    payload: unknown[]
+    appId: string
   }
 }
 
 export default new Service({
   clients,
   routes: {
-    spreadsheetEventBroadcaster: method({
-      POST: [],
+    notify: method({
+      POST: [parseFile, startEventChain],
     }),
   },
 })