+++ node/utils/constants.ts
@@ -1,0 +1,7 @@
+export const APP_ID = 'vtex.spreadsheet-event-broadcaster@0.x'
+
+const _1GB = 1024 ** 3
+
+export const MAXIMUM_FILE_SIZE = _1GB
+
+export const MAXIMUM_FILE_SIZE_STRING = '1GB'