+++ CHANGELOG.md
@@ -6,9 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Notify service route and middlewares
+
 ### Changed
+
 - Setup current linter and prettier presets.
 
 ## [0.0.1] - 2022-01-28
-### Added 
+
+### Added
+
 - Basic setup.